Image Classification/Makerere Fall Armyworm Crop Challenge/#1 Ishtos/requirements.txt
albumentations==1.2.0
grad-cam
hydra-core
matplotlib
numpy
opencv-python
pandas
pre-commit
pytorch-lightning
scikit-learn
timm
# torch
# torchaudio
# torchvision
tqdm
transformers
wandb
warmup_scheduler
